feat: export specific hook
